Add tests for remember_me plugin and fix guard clause inversion
Great example of why unit tests are incredibly useful! Caught that the guard
clause for `forget_me!` was improperly inverted due to going off of the method
description that said we only clear when global persistance is _true_. (the
opposite is the actual intended behavior)

Both the code and the method documentation have been updated to reflect the
intended behavior. (this only affected the rework, when I updated the styling
to use a guard clause and mixed up the conditional)